Move shared mapping seam from models/_mapping.py to utils/_mapping.py
